restructure how the book detail retrievals happen

+++ engine/bookRetrieval.py
@@ -3,42 +3,41 @@ from bs4 import BeautifulSoup
 import re
 import json
 import os
+import time
 
-url = f'https://www.bookreporter.com/coming-soon'
-response = requests.get(url)
-soup = BeautifulSoup(response.content, 'html.parser')
+class LinkManager: 
+    def __init__(self): 
+        self.links = []
+        self.details = []
 
-linksTemp = soup.find_all('div', class_ = 'book-info')
+def LMnger(): 
+    return LinkManager() 
 
-i = 1
-while i < 6:
-    url = f'https://www.bookreporter.com/coming-soon?page={i}'
+def book_links():
+    url = f'https://www.bookreporter.com/coming-soon'
     response = requests.get(url)
     soup = BeautifulSoup(response.content, 'html.parser')
-    temp2 = soup.find_all('div', class_ = 'book-info')
-    for tem in temp2:
-        linksTemp.append(tem)
-    i += 1
 
-links = [] * len(linksTemp)
+    linksTemp = soup.find_all('div', class_ = 'book-info')
 
-for linkie in linksTemp:
-    linkTemp = linkie.find('a')
-    link = linkTemp['href']
-    links.append('https://www.bookreporter.com' + link)
+    i = 1
+    while i < 6:
+        url = f'https://www.bookreporter.com/coming-soon?page={i}'
+        response = requests.get(url)
+        soup = BeautifulSoup(response.content, 'html.parser')
+        temp2 = soup.find_all('div', class_ = 'book-info')
+        for tem in temp2:
+            linksTemp.append(tem)
+        i += 1
 
-bookDets = [""] * (len(links))
+    links = [] * len(linksTemp)
 
-# Data to be written
-dictionary = {
-    "name": "sathiyajith",
-    "rollno": 56,
-    "cgpa": 8.6,
-    "phonenumber": "9976770500"
-}
- 
-# Serializing json
-json_object = None
+    for linkie in linksTemp:
+        linkTemp = linkie.find('a')
+        link = linkTemp['href']
+        links.append('https://www.bookreporter.com' + link)
+    
+    return links
 
 
 def get_book(book_url):
@@ -65,34 +64,41 @@ def get_book(book_url):
 
     return {'title' : title, 'author' : author, 'genres' : category, 'review' : review}
 
-i = 0
+def first_retrieval(linkFile, bookDetsFile):
+    i = 0
+    while i < 30:
+        bookDetsFile[i] = get_book(linkFile[i])
+        i += 1
 
-if os.path.exists("../data/data.json"):
-  os.remove("../data/data.json")
+    if os.path.exists("../data/data.json"):
+        os.remove("../data/data.json")
+    file = open('../data/data.json','w')
+    
+    json.dump(bookDetsFile, file, indent = 4)
 
-file = open('../data/data.json','w')
-'''
-def write_json(new_data):
-    with open('../data/data.json','r+') as file:
-
-        # First we load existing data into a dict.
-        if len(file) == 0:
-            file_data = json.load(file)
-            # Join new_data with file_data inside emp_details
-            file_data.append(new_data)
-            # Sets file's current position at offset.
-            file.seek(0)
-            # convert back to json.
-            json.dump(file_data, file, indent = 4)
-        else:
-            json.dump(new_data, indent = 4)
-'''
-for link in links:
-    bookDets[i] = get_book(link)
-    i += 1
+    return bookDetsFile
 
 
+def retrieve_more(linkFile, bookDetsFile, ind):
+    startInd = 30 * ind
+    endInd = 30 * (ind + 1)
 
-json.dump(bookDets, file, indent = 4)
+    if startInd >= len(linkFile):
+        return bookDetsFile
+    
+    while startInd < endInd:
+        bookDetsFile[startInd] = get_book(linkFile[startInd])
+        startInd += 1
+    
+    print(startInd)
+    print(endInd)
 
- 
+    if os.path.exists("../data/data.json"):
+        os.remove("../data/data.json")
+    file = open('../data/data.json','w')
+    
+    json.dump(bookDetsFile, file, indent = 4)
+
+    return bookDetsFile
+
+
